Only symlink public stuff if the project has a `public` directory

+++ zsh/open-source.zsh
@@ -46,19 +46,24 @@ osslink() {
 
     # public/vendor symlink
     if [[ $PACAKGE_VENDOR == "statamic" && $PACKAGE_NAME == "cms" ]]; then
-        rm -rf public/vendor/statamic
-
-        mkdir -p public/vendor/statamic
+        if [ -d "public" ]; then
+            rm -rf public/vendor/statamic
+            mkdir -p public/vendor/statamic
+        fi
 
         ln -s $CLONE_LOCATION/resources/dist public/vendor/statamic/cp
         ln -s $CLONE_LOCATION/resources/dist-dev public/vendor/statamic/cp-dev
         ln -s $CLONE_LOCATION/resources/dist-frontend public/vendor/statamic/frontend
         ln -s $CLONE_LOCATION/resources/js/package $CLONE_LOCATION/resources/dist-package
     elif [[ -d "$CLONE_LOCATION/dist/build" ]]; then
-        rm -rf public/vendor/$PACKAGE_NAME
-        ln -s $CLONE_LOCATION/dist public/vendor/$PACKAGE_NAME
+        if [ -d "public" ]; then
+            rm -rf public/vendor/$PACKAGE_NAME
+            ln -s $CLONE_LOCATION/dist public/vendor/$PACKAGE_NAME
+        fi
     elif [[ -d "$CLONE_LOCATION/resources/dist/build" ]]; then
-        rm -rf public/vendor/$PACKAGE_NAME
-        ln -s $CLONE_LOCATION/resources/dist public/vendor/$PACKAGE_NAME
+        if [ -d "public" ]; then
+            rm -rf public/vendor/$PACKAGE_NAME
+            ln -s $CLONE_LOCATION/resources/dist public/vendor/$PACKAGE_NAME
+        fi
     fi
 }
